Fixed on load blink issue and onload function call
