+++ common/journey-utils.js
@@ -1,20 +1,11 @@
 /* eslint no-bitwise: ["error", { "allow": ["^", ">>", "&"] }] */
 
 import {
-  santizedFormDataWithContext,
-  urlPath,
   generateUUID,
 } from './formutils.js';
-import { fetchJsonResponse } from './makeRestAPI.js';
-
-import * as CONSTANT from './constants.js';
-import * as CC_CONSTANT from '../creditcards/corporate-creditcard/constant.js';
-
-const { ENDPOINTS, CHANNEL, CURRENT_FORM_CONTEXT: currentFormContext } = CONSTANT;
-const { JOURNEY_NAME } = CC_CONSTANT;
 
 /**
-   * generates the journeyId
+   * @name createJourneyId
    * @param {string} visitMode - The visit mode (e.g., "online", "offline").
    * @param {string} journeyAbbreviation - The abbreviation for the journey.
    * @param {string} channel - The channel through which the journey is initiated.
@@ -25,123 +16,6 @@ function createJourneyId(visitMode, journeyAbbreviation, channel, globals) {
   // var dispInstance = getDispatcherInstance();
   const journeyId = globals.functions.exportData().journeyId || `${dynamicUUID}_01_${journeyAbbreviation}_${visitMode}_${channel}`;
   globals.functions.setProperty(globals.form.runtime.journeyId, { value: journeyId });
-  return journeyId;
 }
 
-const getCurrentContext = () => currentFormContext;
-
-/**
-   * @name invokeJourneyDropOff to log on success and error call backs of api calls
-   * @param {state} state
-   * @param {string} mobileNumber
-   * @param {Object} globals - globals variables object containing form configurations.
-   * @return {PROMISE}
-   */
-const invokeJourneyDropOff = async (state, mobileNumber, globals) => {
-  const DEFAULT_MOBILENO = '9999999999';
-  const journeyJSONObj = {
-    RequestPayload: {
-      userAgent: (typeof window !== 'undefined') ? window.navigator.userAgent : 'onLoad',
-      leadProfile: {
-        mobileNumber: mobileNumber || DEFAULT_MOBILENO,
-      },
-      formData: {
-        channel: CHANNEL,
-        journeyName: JOURNEY_NAME,
-        journeyID: globals.form.runtime.journeyId.$value || createJourneyId('online', JOURNEY_NAME, CHANNEL, globals),
-        journeyStateInfo: [
-          {
-            state,
-            stateInfo: JSON.stringify(santizedFormDataWithContext(globals)),
-            timeinfo: new Date().toISOString(),
-          },
-        ],
-      },
-    },
-  };
-  const url = urlPath(ENDPOINTS.journeyDropOff);
-  const method = 'POST';
-  // return globals.functions.exportData().queryParams.leadId ? fetchJsonResponse(url, journeyJSONObj, method) : null;
-  return journeyJSONObj.RequestPayload.formData.journeyID ? fetchJsonResponse(url, journeyJSONObj, method) : null;
-};
-
-/**
-   * @name invokeJourneyDropOffUpdate
-   * @param {string} state
-   * @param {string} mobileNumber
-   * @param {string} leadProfileId
-   * @param {string} journeyId
-   * @param {Object} globals - globals variables object containing form configurations.
-   * @return {PROMISE}
-   */
-const invokeJourneyDropOffUpdate = async (state, mobileNumber, leadProfileId, journeyId, globals) => {
-  const sanitizedFormData = santizedFormDataWithContext(globals, currentFormContext);
-  const journeyJSONObj = {
-    RequestPayload: {
-      userAgent: (typeof window !== 'undefined') ? window.navigator.userAgent : '',
-      leadProfile: {
-        mobileNumber,
-        leadProfileId: leadProfileId?.toString(),
-      },
-      formData: {
-        channel: CHANNEL,
-        journeyName: JOURNEY_NAME,
-        journeyID: journeyId,
-        journeyStateInfo: [
-          {
-            state,
-            stateInfo: JSON.stringify(sanitizedFormData),
-            timeinfo: new Date().toISOString(),
-          },
-        ],
-      },
-    },
-  };
-  const url = urlPath(ENDPOINTS.journeyDropOffUpdate);
-  const method = 'POST';
-  return fetchJsonResponse(url, journeyJSONObj, method);
-};
-
-/**
-   * @name printPayload
-   * @param {string} payload.
-   * @param {object} formContext.
-   * @returns {object} currentFormContext.
-   */
-function journeyResponseHandlerUtil(payload, formContext) {
-  formContext.leadProfile = {};
-  formContext.leadProfile.leadProfileId = String(payload);
-  return formContext;
-}
-
-/**
-  * @name invokeJourneyDropOffByParam
-  * @param {string} mobileNumber
-  * @param {string} leadProfileId
-  * @param {string} journeyId
-  * @return {PROMISE}
-  */
-const invokeJourneyDropOffByParam = async (mobileNumber, leadProfileId, journeyID) => {
-  const journeyJSONObj = {
-    RequestPayload: {
-      leadProfile: {
-        mobileNumber,
-      },
-      journeyInfo: {
-        journeyID,
-      },
-    },
-  };
-  const url = urlPath(ENDPOINTS.journeyDropOffParam);
-  const method = 'POST';
-  return fetchJsonResponse(url, journeyJSONObj, method);
-};
-
-export {
-  invokeJourneyDropOff,
-  invokeJourneyDropOffByParam,
-  invokeJourneyDropOffUpdate,
-  journeyResponseHandlerUtil,
-  getCurrentContext,
-  createJourneyId,
-};
+export default createJourneyId;
