+++ creditcards/fd-card/constant.js
@@ -21,6 +21,7 @@ const ERROR_MSG = {
 const FD_ENDPOINTS = {
   otpGen: '/content/hdfc_hafcards/api/customeridentificationotpgen.json',
   otpVal: '/content/hdfc_hafcards/api/otpvalidationandcardsinquiry.json',
+  journeyDropOff: '/content/hdfc_commonforms/api/journeydropoff.json',
 };
 
 const OTP_TIMER = 30;
