+++ creditcards/fd-card/fd-functions.js
@@ -7,18 +7,13 @@ import {
   resendOTP,
   customSetFocus,
   reloadPage,
-  onPageLoad,
 } from './fdlien-functions.js';
+import invokeJourneyDropOff from './fd-journey-util.js';
+
 // import { getOTP, otpValidation } from '../../common/functions.js'; // improvisation required to make it generic, till then using the journey specific function getotp,otpVal 👆
 import { redirect } from '../../common/functions.js';
-import {
-  createJourneyId,
-  getCurrentContext,
-  invokeJourneyDropOff,
-  invokeJourneyDropOffByParam,
-  invokeJourneyDropOffUpdate,
-  journeyResponseHandlerUtil,
-} from '../../common/journey-utils.js';
+
+import createJourneyId from '../../common/journey-utils.js';
 
 import {
   addGaps,
@@ -31,11 +26,6 @@ export {
   otpValidation,
   validateLogin,
   createJourneyId,
-  getCurrentContext,
-  invokeJourneyDropOff,
-  invokeJourneyDropOffByParam,
-  invokeJourneyDropOffUpdate,
-  journeyResponseHandlerUtil,
   maskedMobNum,
   addGaps,
   addMobileValidation,
@@ -43,5 +33,5 @@ export {
   resendOTP,
   customSetFocus,
   reloadPage,
-  onPageLoad,
+  invokeJourneyDropOff,
 };
